Feat: Upgrade to BepInEx
Switch from injecting via QMod Manager Reloaded to BepInEx.  This
increases modding capabilities and improves OS support.

Update the TargetFramework per BepInEx's guidelines.

Reference BepInEx assemblies instead of Harmony since it is wrapped by
them.  Exclude BepInEx assemblies from merging.

Generate the BepInEx Plugin based on assembly attributes.

The standard approach is to generate constants representing the
assembly attributes and manually write a Plugin class that uses these.
This would require a Plugin for each separate mod.  Attempting to move
this into a lib does not work since it will be compiled with the lib's
attributes.  By generating the Plugin itself, this can safely be moved
into a lib.  The generated code is not significantly more complicated
than just generating constants since it points to the MainPatcher.

If more complicated BepInEx attributes are needed, it should be
possible to inherit from the generated Plugin.  This is unlikely to be
necessary in this solution.