Feat: Cap the maximum number of footprints
Destroy the oldest footprint (trail) when the total count exceeds a
threshold.

Huge numbers of trails can accumulate in one area, dropping performance
excessively even on high powered desktops.  On weaker devices (e.g.,
Steam Deck), this impact is even more significant.  I believe the
performance degradation is due to the lack of batching, but this
appears to be impossible to alter after Unity is built.

Removing trails that are visible to the camera is jarring.  This
implementation waits until they are no longer visible before destroying
to minimize the visual impact.