Feat: Enable extension of generated Plugin
Make the generated Plugin a partial class so that individual mods can
extend it with additional behavior (e.g., configs).

Inheriting from Plugin does not work because BepInEx creates an object
of the type the annotation is attached to and not of any children.
This means that overrides or Unity calls are not invoked and it cannot
be extended.

Alternate approaches could include manually writing Plugin for all mods
or reflection for a custom OnLoad attribute.